newpins >> pins
